Improve logging tags in worker's sync processes
